Implement number literals and `+` keyword
This commit adds support for number literals and the `+` keyword in Stacko source code.

When number literals are encountered they are pushed to the top of the stack for later use.

The `+` keyword takes the two top-most values on the stack, adds their values together, and then pushes the result of this operation onto the top of the stack. This entire operating is type-checked at run-time.

Co-Authored-By: Suirabu <suirabu.dev@gmail.com>