Added syntax highlighting support for VS Code
Added complete syntax highlighting support for VS Code.

This can be built as an extension using the following commands:

```
$ npm install
$ npm install vsce -g
$ vsce package
```

This can then be used in VS Code by switching to the Extensions view on
the left window, pressing the elipsis icon at the top right, selecting
"Install from VSIX" from the drop-down menu and selecting the file you
just created.